third_party/blink/web_tests/external/wpt/mathml/presentation-markup/operators/mo-axis-height-1-expected.txt

This is a testharness.js-based test.
FAIL AxisHeight (size variant) assert_approx_equals: mo: size expected 140 +/- 5 but got 11
FAIL AxisHeight (glyph assembly) assert_approx_equals: mo: size expected 300 +/- 5 but got 11
Harness: the test ran to completion.
